+++ lib/oxaiworkers/assistant/module_base.rb
@@ -5,7 +5,7 @@ module OxAiWorkers
     module ModuleBase
       include OxAiWorkers::LoadI18n
 
-      attr_accessor :iterator
+      attr_accessor :iterator, :role, :description, :capabilities, :id, :title
 
       def task=(task)
         @iterator.cleanup
@@ -24,6 +24,11 @@ module OxAiWorkers
         worker = delayed ? DelayedRequest.new : Request.new(on_stream:)
         worker.model = model || OxAiWorkers.configuration.model
         worker
+      end
+
+      def replace_context(context)
+        @iterator.clear_context
+        @iterator.add_context context
       end
     end
   end